feat auto update sdk

+++ volc-sdk-java/src/main/java/com/volcengine/model/request/ApplySmsSignatureRequest.java
@@ -41,5 +41,5 @@ public class ApplySmsSignatureRequest {
     String from;
 
     @JSONField(name = "signatureIdentificationID")
-    int signatureIdentificationID;
+    long signatureIdentificationID;
 }
